refactor(git): introduce GitBackend trait + CliBackend
Phase 1 of the gix migration: extract the existing shell-out git logic
from src/utils.rs into a new src/git/ module behind a GitBackend trait,
with the CLI shell-out implementation as the only backend. operations/
callers now go through crate::git::{clone_repo,update_repo,is_working_dir_clean}
instead of the utils::git_* functions directly.

Pure restructuring: function bodies are byte-identical to the originals.
No behavior change. Sets up Phase 2 to add a gix-based backend alongside.
